Move body of vk_instance to separate file

+++ testing.cpp
@@ -7,6 +7,7 @@
 #include <vector>
 #include <stdexcept>
 #include <optional>
+#include <unordered_set>
 
 const uint32_t INIT_WIDTH = 800, INIT_HEIGHT = 600;
 
@@ -45,9 +46,13 @@ int main() {
 	auto glfw_window = GWindow(INIT_WIDTH, INIT_HEIGHT);
 
 	VkInstance instance;
-	// Will only actually be filled if validation is enabled
-	VkDebugUtilsMessengerEXT debug_msgr;
+	// Will only actually be set if validation is enabled
+	VkDebugUtilsMessengerEXT debug_msgr{};
 	vk_instance::create(glfw_window.req_instance_exts, {}, VALIDATION_ENABLED, &instance, &debug_msgr);
+
+	// Create surface
+	VkSurfaceKHR surface;
+	if (glfwCreateWindowSurface(instance, glfw_window.window, nullptr, &surface) != VK_SUCCESS) throw std::runtime_error("Could not create surface!");
 
 	// Create physical device
 	uint32_t phys_dev_ct;
@@ -61,41 +66,60 @@ int main() {
 	vkGetPhysicalDeviceProperties(phys_dev, &phys_dev_props);
 	std::cout << "Using device: " << phys_dev_props.deviceName << std::endl;
 
-	// Find queue family
+	// Find queue families
 	std::optional<uint32_t> graphics_fam;
+	std::optional<uint32_t> present_fam;
 	uint32_t queue_fam_ct;
 	vkGetPhysicalDeviceQueueFamilyProperties(phys_dev, &queue_fam_ct, nullptr);
 	std::vector<VkQueueFamilyProperties> queue_fams(queue_fam_ct);
 	vkGetPhysicalDeviceQueueFamilyProperties(phys_dev, &queue_fam_ct, queue_fams.data());
 
-	for (size_t i = 0; i < queue_fam_ct; ++i) {
-		if (queue_fams[i].queueFlags & VK_QUEUE_GRAPHICS_BIT) {
-			graphics_fam = i;
-			break;
-		}
+	for (uint32_t i = 0; i < queue_fam_ct; ++i) {
+		if (queue_fams[i].queueFlags & VK_QUEUE_GRAPHICS_BIT) graphics_fam = i;
+		VkBool32 present_supported;
+		vkGetPhysicalDeviceSurfaceSupportKHR(phys_dev, i, surface, &present_supported);
+		if (present_supported) present_fam = i;
+
+		if (graphics_fam.has_value() && present_fam.has_value()) break;
 	}
 
-	if (!graphics_fam.has_value()) throw std::runtime_error("Could not find queue family with graphics support!");
+	if (!graphics_fam.has_value() || !present_fam.has_value())
+		throw std::runtime_error("Could not find queue families supporting graphics and presentation!");
+
+	std::unordered_set<uint32_t> unique_queue_fams;
+	unique_queue_fams.insert(graphics_fam.value());
+	unique_queue_fams.insert(present_fam.value());
 
 	// Create logical device
 	VkDevice device;
 
-	VkDeviceQueueCreateInfo dev_queue_info{};
-	dev_queue_info.sType = VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO;
-	dev_queue_info.queueCount = 1;
-	dev_queue_info.queueFamilyIndex = graphics_fam.value();
+	// Necessary because the device has to support all our different queue
+	// families
+	std::vector<VkDeviceQueueCreateInfo> dev_queue_infos;
 	float queue_priority = 1.0f;
-	dev_queue_info.pQueuePriorities = &queue_priority;
+	for (auto queue_fam : unique_queue_fams) {
+		VkDeviceQueueCreateInfo dev_queue_info{};
+		dev_queue_info.sType = VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO;
+		dev_queue_info.queueCount = 1;
+		dev_queue_info.queueFamilyIndex = queue_fam;
+		dev_queue_info.pQueuePriorities = &queue_priority;
+		dev_queue_infos.push_back(dev_queue_info);
+        }
 
-	VkPhysicalDeviceFeatures dev_features{};
+        VkPhysicalDeviceFeatures dev_features{};
 
 	VkDeviceCreateInfo device_info{};
 	device_info.sType = VK_STRUCTURE_TYPE_DEVICE_CREATE_INFO;
-	device_info.queueCreateInfoCount = 1;
-	device_info.pQueueCreateInfos = &dev_queue_info;
+	device_info.queueCreateInfoCount = static_cast<uint32_t>(dev_queue_infos.size());
+	device_info.pQueueCreateInfos = dev_queue_infos.data();
 	device_info.pEnabledFeatures = &dev_features;
 
 	if (vkCreateDevice(phys_dev, &device_info, nullptr, &device) != VK_SUCCESS) throw std::runtime_error("Could not create device!");
+
+	// Create queues
+	VkQueue graphics_queue, present_queue;
+	vkGetDeviceQueue(device, graphics_fam.value(), 0, &graphics_queue);
+	vkGetDeviceQueue(device, present_fam.value(), 0, &present_queue);
 
 	while (!glfwWindowShouldClose(glfw_window.window)) {
 		glfwPollEvents();
@@ -103,6 +127,8 @@ int main() {
 
 	vkDestroyDevice(device, nullptr);
 
-	vk_instance::destroy_debug_msgr(instance, debug_msgr);
+	vkDestroySurfaceKHR(instance, surface, nullptr);
+
+	if (debug_msgr) vk_instance::destroy_debug_msgr(instance, debug_msgr);
 	vkDestroyInstance(instance, nullptr);
 }
